opt type

+++ cell-specification/src/main/java/com/lazycece/cell/specification/model/Cell.java
@@ -45,7 +45,7 @@ public class Cell {
     /**
      * The sequence
      */
-    private Long sequence;
+    private Integer sequence;
 
     public String getCode() {
         return code;
@@ -79,11 +79,11 @@ public class Cell {
         this.machine = machine;
     }
 
-    public Long getSequence() {
+    public Integer getSequence() {
         return sequence;
     }
 
-    public void setSequence(Long sequence) {
+    public void setSequence(Integer sequence) {
         this.sequence = sequence;
     }
